cleaned up code, and moved test
